+++ circuits/cpp/barretenberg/src/aztec/numeric/uint256/uint256_impl.hpp
@@ -146,9 +146,9 @@ constexpr std::pair<uint256_t, uint256_t> uint256_t::mul_extended(const uint256_
     return { lo, hi };
 }
 
-constexpr uint256_t uint256_t::slice(const size_t start, const size_t end) const
+constexpr uint256_t uint256_t::slice(const uint64_t start, const uint64_t end) const
 {
-    const size_t range = end - start;
+    const uint64_t range = end - start;
     const uint256_t mask = (range == 256) ? -uint256_t(1) : (uint256_t(1) << range) - 1;
     return ((*this) >> start) & mask;
 }
